+++ NEWS.md
@@ -1,6 +1,7 @@
 # `rsample` 0.0.2
 
 * `initial_split`, `training`, and `testing` were added to do training/testing splits prior to resampling. 
+* `caret2rsample` and `rsample2caret` can convert `rset` objects to those used by `caret::trainControl` and vice-versa. 
 
 # `rsample` 0.0.1